requirements.txt
snap-stanford
matplotlib
networkx
numpy